@@ -1,20 +1,21 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-//apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: "androidx.navigation.safeargs.kotlin"
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion rootProject.ext.buildToolsVersion
+    compileSdkVersion 29
+
     defaultConfig {
         applicationId "io.github.myolwin00.sunshine"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
-    	buildConfigField 'String', 'WeatherApiKey', WeatherApiKey
+        buildConfigField 'String', 'WeatherApiKey', WeatherApiKey
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
