@@ -1,12 +1,9 @@
 package io.github.myolwin00.sunshine.data.viewmodel;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.arch.lifecycle.ViewModelProviders;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
+import androidx.lifecycle.ViewModelProvider;
 
 import io.github.myolwin00.sunshine.data.repository.WeatherRepository;
-import io.github.myolwin00.sunshine.view.holders.ForecastVH;
 
 /**
  * Created by emrys on 2/19/18.
