@@ -1,8 +1,8 @@
 package io.github.myolwin00.sunshine.data.source.local
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
 import android.content.Context
 import io.github.myolwin00.sunshine.data.Forecast
 import io.github.myolwin00.sunshine.data.Weather
