@@ -1,12 +1,7 @@
 package io.github.myolwin00.sunshine.data.source.local
 
-import android.arch.persistence.room.TypeConverter
-import com.google.gson.Gson
+import androidx.room.TypeConverter
 import io.github.myolwin00.sunshine.data.Weather
-import com.google.gson.reflect.TypeToken
-
-
-
 
 /**
  * Created by myolwin00 on 11/13/17.
@@ -15,11 +10,11 @@ class WeatherTypeConverter {
 
     @TypeConverter
     fun toStr(weathers: List<Weather>): String {
-        return Gson().toJson(weathers)
+        TODO("relation table")
     }
 
     @TypeConverter
     fun toList(weatherStr: String): List<Weather> {
-        return Gson().fromJson(weatherStr, object : TypeToken<List<Weather>>(){}.type)
+        TODO("relation table")
     }
 }
